Added conftest to import the dependancies required for continuous testing Removed redundancy from add_punctuation_transcript
